+++ dbx/models/workflow/v2dot0/parameters.py
@@ -1,8 +1,15 @@
+from pydantic import root_validator
+
+from dbx.models.validators import mutually_exclusive
 from dbx.models.workflow.common.parameters import BaseParametersMixin, ParametersMixin, StandardBasePayload
 
 
-class AssetBasedParametersPayload(BaseParametersMixin, ParametersMixin):
+class AssetBasedRunPayload(BaseParametersMixin, ParametersMixin):
     """"""
+
+    _validate_unique = root_validator(pre=True)(
+        lambda _, values: mutually_exclusive(["parameters", "named_parameters"], values)
+    )
 
 
 class StandardRunPayload(StandardBasePayload):