+++ CHANGELOG.md
@@ -16,7 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - 📖 documentation on the differences between `dbx execute` and `dbx launch`
-- 📖 documentation on how to use parameter passing in various cases.
+- 📖 documentation on how to use parameter passing in various cases
+- 📖 documentation on how to enable Photon
 
 ### Fixed